Exception class for general errors
Some of Exceptions rised in the library was causing confusion when
checking the source of error. An API connection error was being
identified the same way that a credit card validation failure.
Change general and technical errors to rise a base Exception,
making possible to identify when business rules errors, like
payment validations errors, ocurred.